+++ frontend/app/(app)/(tabs)/_layout.tsx
@@ -2,8 +2,8 @@ import { Box } from "@/design-system/base/box";
 import SwitchGroupButton from "@/design-system/components/shared/buttons/select-group";
 import Setting from "@/design-system/components/shared/buttons/setting";
 import { Icon } from "@/design-system/components/shared/icons/icon";
-import SettingPopup from "@/design-system/components/shared/setting-popup";
-import SwitchGroupBottomSheet from "@/design-system/components/shared/switch-group-sheet";
+import SettingPopup from "@/design-system/components/shared/settings/setting-popup";
+import SwitchGroupBottomSheet from "@/design-system/components/shared/switch-group/switch-group-sheet";
 import { useIsBasicMode } from "@/hooks/component/mode";
 import BottomSheet from "@gorhom/bottom-sheet";
 import { router, Tabs } from "expo-router";
